MFC r258437:
Sync with NetBSD.  The functional change is to make the output when
comparing a directory to an mtree file more compatible with fmtree when
FreeBSD 9 compatiblity mode is on.  This output is clearly intended for
humans not computers, but some tools such as mergemaster's -U option rely
on it.

Approved by:	re (glebius)


git-svn-id: https://svn.freebsd.org/base/stable/10@258655 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f