MFC: r258456
Fix mergemaster -U by forcing FreeBSD 9 compatiblity in mtree when mtree is
nmtree.

The mtree output used by mergemaster in this case was clearly not meant for
computer consumption and an approach based on -f <file1> -f <file2> would
probably be a better idea, but this is a minimal change.

Approved by:	re (glebius)


git-svn-id: https://svn.freebsd.org/base/stable/10@258657 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f