Hoist all the mergeinfo up to the root in preparation for enforcing merges to the root only.  All MFC's were rerecorded to the root.
Going forward, if an MFC includes mergeinfo, it will need to be made to
the root and committed from the root.  Merges with --ignore-ancestry
or diff | patch can go anywhere.

The mergeinfo in HEAD is in a bad state from years of neglect and manual
tampering and this was branched into 10.x.  This confuses the coalescing
code and prevents it from doing its job.

Approved by:	re (gjb, implicit)


git-svn-id: https://svn.freebsd.org/base/stable/10@259073 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f