Remove svn:mergeinfo from the releng/10.0 branch.
After branch creation from stable/10, the stable/10 branch mergeinfo
was moved to the root of the branch.

Since there have not been any merges from stable/10 to releng/10.0
yet, we do not need to track any of the existing mergeinfo here.

Merges to releng/10.0 should now be done to the root of the branch.

For future branches during the release cycle, unless otherwise noted,
this change will be done as part of the stable/ and releng/ branch
creation.

Discussed with:	peter
Approved by:	re (implicit)
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/releng/10.0@259128 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f