MFC r241959:
Fix a number of other clang warnings in libstdc++, which could appear
when building other C++ software with it.  Also fix one actual bug in
libsupc++, which was exposed by such a warning.  This latter fix is the
only functional change.

This is required for newer versions of gcc and clang which have
stricter warnings about parenthesis in C++ programs.


git-svn-id: https://svn.freebsd.org/base/stable/8@259263 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f