Modify handling of writes to the vlapic ICR_TIMER, DCR_TIMER, ICRLO and ESR registers.
The handler is now called after the register value is updated in the virtual
APIC page. This will make it easier to handle APIC-write VM-exits with APIC
register virtualization turned on.

We can no longer rely on the value of 'icr_timer' on the APIC page
in the callout handler. With APIC register virtualization the value of
'icr_timer' will be updated by the processor in guest-context before an
APIC-write VM-exit.

Clear the 'delivery status' bit in the ICRLO register in the write handler.
With APIC register virtualization the write happens in guest-context and
we cannot prevent a (buggy) guest from setting this bit.


git-svn-id: https://svn.freebsd.org/base/head@259961 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f